+++ spec/fixtures/dump_lmdb_key.lua
@@ -1,0 +1,4 @@
+local lmdb = require("resty.lmdb")
+local key = assert(arg[1])
+
+ngx.say(lmdb.get(key))